+++ CHANGES.md
@@ -1,13 +1,15 @@
 # node-trace-event changelog
 
-## 1.1.1 (not yet released)
+## 1.2.0 (not yet released)
 
-(nothing yet)
+- Add `trace_event.createBunyanLogger()` usage for some sugar. See the
+  README.md for details.
 
 
 ## 1.1.0
 
-Rename to 'trace-event', which is a much more accurate name.
+- Rename to 'trace-event', which is a much more accurate name.
+
 
 ## 1.0.0